fixing warnigns refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@25879 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netimport/NIImporter_OpenStreetMap.cpp
@@ -1270,7 +1270,7 @@ NIImporter_OpenStreetMap::RelationHandler::myEndElement(int element) {
                                             + "' is corrupt. Probably OSM file is incomplete.");
                             continue;
                         }
-                        NBPTPlatform platform(p[p.size()/2],p.length());
+                        NBPTPlatform platform(p[(int)p.size()/2],p.length());
                         ptStop->addPlatformCand(platform);
                     } else {
                         if (myOSMNodes.find(myPlatform.ref) == myOSMNodes.end()) {
